Environment conditions generation logic, support for later database storage

+++ DataCollectionService/Data/Models/Humidity.cs
@@ -2,5 +2,5 @@ namespace DataCollectionService.Business.Models;
 
 public class Humidity : ModelBase
 {
-    public decimal VaporGramsPerAirKg { get; set; }
+    public double VaporGramsPerAirKg { get; set; }
 }
